+++ CHANGES.md
@@ -8,6 +8,9 @@
 - Search button on HoveringToolBar  
 - Insert Local Image  
 
+### Fixed
+- TODO: normalize paths for Window, Linux, macOS
+
 ## [app-v0.1.0] - 2022-03-03
 
 initial release